+++ svelte.config.js
@@ -1,9 +1,11 @@
 import adapter from '@sveltejs/adapter-cloudflare';
+import { mdsvex } from 'mdsvex';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	preprocess: vitePreprocess(),
+	preprocess: [vitePreprocess(), mdsvex({ extensions: ['.md'] })],
+	extensions: ['.svelte', '.md'],
 
 	kit: {
 		adapter: adapter()